fix ErrorModal.ts
